microservices/attendance-service/requirements.txt
Django==4.2.7
djangorestframework==3.14.0
django-cors-headers==4.3.1
drf-spectacular==0.26.5
django-filter==23.3
psycopg2-binary==2.9.7
django-redis==5.4.0
python-decouple==3.8
gunicorn==21.2.0
celery==5.3.4
